1. Move Implementation from core to domain 2. IMappedAccessVerifier, mapping a resourceEntity to another 3. Add a bean: SecurityVerifierAgent.COMPONENT_NAME)

+++ general/general-dynamic-dataservice/src/main/java/com/taoswork/tallybook/dynamic/dataservice/core/persistence/impl/PersistenceManagerImpl.java
@@ -5,6 +5,9 @@ import com.taoswork.tallybook.dynamic.dataservice.core.metaaccess.DynamicEntityM
 import com.taoswork.tallybook.dynamic.dataservice.core.persistence.PersistenceManager;
 import com.taoswork.tallybook.dynamic.dataservice.core.query.dto.CriteriaQueryResult;
 import com.taoswork.tallybook.dynamic.dataservice.core.query.dto.CriteriaTransferObject;
+import com.taoswork.tallybook.dynamic.dataservice.core.security.ISecurityVerifier;
+import com.taoswork.tallybook.dynamic.dataservice.core.security.impl.SecurityVerifierAgent;
+import com.taoswork.tallybook.general.authority.core.basic.Access;
 
 import javax.annotation.Resource;
 
@@ -19,29 +22,37 @@ public class PersistenceManagerImpl implements PersistenceManager {
     @Resource(name = DynamicEntityMetadataAccess.COMPONENT_NAME)
     protected DynamicEntityMetadataAccess dynamicEntityMetadataAccess;
 
+    @Resource(name = SecurityVerifierAgent.COMPONENT_NAME)
+    protected ISecurityVerifier securityVerifier;
+
     @Override
     public <T> T persist(T entity){
+        securityVerifier.checkAccess(entity.getClass().getName(), Access.Read, entity);
         return dynamicEntityDao.persist(entity);
     }
 
     @Override
     public <T> T find(Class<T> entityClz, Object key) {
+        securityVerifier.checkAccess(entityClz.getName(), Access.Read);
         Class<T> entityRootClz = this.dynamicEntityMetadataAccess.getRootInstanceableEntityClass(entityClz);
         return dynamicEntityDao.find(entityRootClz, key);
     }
 
     @Override
     public <T> T update(T entity){
+        securityVerifier.checkAccess(entity.getClass().getName(), Access.Update, entity);
         return dynamicEntityDao.update(entity);
     }
 
     @Override
     public <T> void delete(T entity){
+        securityVerifier.checkAccess(entity.getClass().getName(), Access.Delete, entity);
         dynamicEntityDao.remove(entity);
     }
 
     @Override
     public <T> CriteriaQueryResult<T> query(Class<T> entityClz, CriteriaTransferObject query){
+        securityVerifier.checkAccess(entityClz.getName(), Access.Query);
         Class<T> entityRootClz = this.dynamicEntityMetadataAccess.getRootInstanceableEntityClass(entityClz);
         return dynamicEntityDao.query(entityRootClz, query);
     }
